+++ src/aurite/components/llm/providers/openai_client.py
@@ -22,8 +22,8 @@ from ..base_client import (
     DEFAULT_TEMPERATURE as BASE_DEFAULT_TEMPERATURE,
     # DEFAULT_SYSTEM_PROMPT is also available if needed
 )
-from ...components.agents.agent_models import AgentOutputMessage, AgentOutputContentBlock
-from ...config.config_models import LLMConfig
+from ...agents.agent_models import AgentOutputMessage, AgentOutputContentBlock
+from ....config.config_models import LLMConfig
 
 logger = logging.getLogger(__name__)
